+++ frontend/src/models/Coordinator.model.ts
@@ -307,9 +307,9 @@ export class Coordinator {
   getRelayUrl = (network: 'mainnet' | 'testnet', hostUrl: string, selfHosted: boolean): string => {
     const protocol = hostUrl.includes('https') ? 'wss' : 'ws';
     if (selfHosted && this.shortAlias !== 'local') {
-      return `${protocol}://${hostUrl.replace(/^https?:\/\//, '')}/${network}/${this.shortAlias}/relay/`;
+      return `${protocol}://${hostUrl.replace(/^https?:\/\//, '')}/${network}/${this.shortAlias}/nostr/`;
     } else {
-      return `${protocol}://${this.url.replace(/^https?:\/\//, '')}/relay/`;
+      return `${protocol}://${this.url.replace(/^https?:\/\//, '')}/nostr/`;
     }
   };
 }
